Change to floating point z values, fix sort order, tame isometric axes.
Rewrote the sort to actually do what it was supposed to. Originally rewrote
iso.go but figured it was pretty arbitrary no matter which way the axes went.
For reference, dear commit reader, x increases towards the top right, y
increases towards the bottom right, and z increases upwards from the x-y plane.
Anyway, fixed the sort, changed to the discussed variable-length z slices,
and added a 3D function to graph with the tiles.